Add minor handling for CTRL-d and errors Fixed an issue where an empty command line would result in a segmention fault Fixed an issue where execute command would not close child proccess on failure
