+++ execute_command.c
@@ -9,20 +9,22 @@
 void execute_command(char *pathname, char **arg_array)
 {
 	int status;
-/*	pid_t child_pid = fork();
- */
+	char *error[] = {"/bin/echo" , "-n", "Error\n", NULL};
+
 	if (fork() == 0)
 	{
 		/* execute command @ pathname w/ arguments in arg_array *
 		 * or print an error on failure                         */
 		if ((execve(pathname, arg_array, NULL)) == -1)
 		{
-			printf("Error\n");
-
+			/* will print "error" using echo to close pid */
+			execve(error[0], error, NULL);
 		}
 	}
 	else
 	{
 		wait(&status);
 	}
+
+	return;
 }
